+++ src/Particle/ParticleSpatialLayout.hpp
@@ -127,7 +127,7 @@ namespace ippl {
         Kokkos::parallel_for(
             "set invalid",
             localnum,
-            KOKKOS_LAMBDA(const int i) {
+            KOKKOS_LAMBDA(const size_t i) {
                 if (invalid(i)) {
                     pdata.ID(i) = -1;
                 }
@@ -194,7 +194,7 @@ namespace ippl {
         Kokkos::parallel_scan(
             "ParticleSpatialLayout::fillHash()",
             ranks.extent(0),
-            KOKKOS_LAMBDA(const int i, int& idx, const bool final) {
+            KOKKOS_LAMBDA(const size_t i, int& idx, const bool final) {
                 if (final) {
                     if (rank == ranks(i)) {
                         hash(idx) = i;
